Add camera focus function (#354)
* add camera focus function

* use camera focus function on thumnail
