+++ packages/spatial/src/camera/functions/CameraFunctions.ts
@@ -23,11 +23,15 @@ All portions of the code written by the Infinite Reality Engine team are Copyrig
 Infinite Reality Engine. All Rights Reserved.
 */
 
-import { ComponentType, getOptionalComponent, setComponent } from '@ir-engine/ecs/src/ComponentFunctions'
+import { ComponentType, getComponent, getOptionalComponent, setComponent } from '@ir-engine/ecs/src/ComponentFunctions'
 import { Entity } from '@ir-engine/ecs/src/Entity'
 
-import { Box3, PerspectiveCamera, Sphere, Vector3 } from 'three'
+import { Box3, Matrix4, PerspectiveCamera, Quaternion, Sphere, Vector3 } from 'three'
+import { BoundingBoxComponent, updateBoundingBox } from '../../transform/components/BoundingBoxComponents'
+import { TransformComponent } from '../../transform/components/TransformComponent'
 import { getBoundingBoxVertices } from '../../transform/functions/BoundingBoxFunctions'
+import { computeTransformMatrix } from '../../transform/systems/TransformSystem'
+import { CameraComponent } from '../components/CameraComponent'
 import { TargetCameraRotationComponent } from '../components/TargetCameraRotationComponent'
 
 export const setTargetCameraRotation = (entity: Entity, phi: number, theta: number, time = 0.3) => {
@@ -84,4 +88,54 @@ export function computeCameraDistanceAndCenter(
 export function computeCameraDistanceAndCenterFromBox(camera: PerspectiveCamera, box: Box3, padding: number = 1.1) {
   const points = getBoundingBoxVertices(box)
   return computeCameraDistanceAndCenter(camera, points, padding)
+}
+export function setCameraFocusOnBox(modelEntity: Entity, cameraEntity: Entity) {
+  updateBoundingBox(modelEntity)
+
+  const bbox = getComponent(modelEntity, BoundingBoxComponent).box
+  const center = bbox.getCenter(new Vector3())
+
+  // Calculate the bounding sphere radius
+  const boundingSphere = bbox.getBoundingSphere(new Sphere())
+  const radius = boundingSphere.radius
+
+  const camera = getComponent(cameraEntity, CameraComponent)
+  const fov = camera.fov * (Math.PI / 180) // convert vertical fov to radians
+
+  // Calculate the camera direction vector with the desired angle offsets
+  const angleY = 30 * (Math.PI / 180) // 30 degrees in radians
+  const angleX = 15 * (Math.PI / 180) // 15 degrees in radians
+
+  const direction = new Vector3(
+    Math.sin(angleY) * Math.cos(angleX),
+    Math.sin(angleX),
+    Math.cos(angleY) * Math.cos(angleX)
+  ).normalize()
+
+  // Calculate the distance from the camera to the bounding sphere such that it fully frames the content
+  const distance = radius / Math.sin(fov / 2)
+
+  // Calculate the camera position
+  const cameraPosition = direction.multiplyScalar(distance).add(center)
+
+  // Set the camera transform component
+  setComponent(cameraEntity, TransformComponent, { position: cameraPosition })
+  computeTransformMatrix(cameraEntity)
+
+  // Calculate the quaternion rotation to look at the center
+  const lookAtMatrix = new Matrix4()
+  lookAtMatrix.lookAt(cameraPosition, center, new Vector3(0, 1, 0))
+  const targetRotation = new Quaternion().setFromRotationMatrix(lookAtMatrix)
+
+  // Apply the rotation to the camera's TransfortexturemComponent
+  setComponent(cameraEntity, TransformComponent, { rotation: targetRotation })
+  computeTransformMatrix(cameraEntity)
+  camera.matrixWorldInverse.copy(camera.matrixWorld).invert()
+
+  // Update the view camera matrices
+  const viewCamera = camera.cameras[0]
+  viewCamera.matrixWorld.copy(camera.matrixWorld)
+  viewCamera.matrixWorldInverse.copy(camera.matrixWorldInverse)
+  viewCamera.projectionMatrix.copy(camera.projectionMatrix)
+  viewCamera.projectionMatrixInverse.copy(camera.projectionMatrixInverse)
 }
